fix: rename schemas files to schema
