+++ app/server/routers/user.ts
@@ -2,7 +2,7 @@ import { ORPCError } from '@orpc/client';
 import { Prisma } from '@prisma/client';
 import { z } from 'zod';
 
-import { zSession, zUser } from '@/features/user/schemas';
+import { zSession, zUser } from '@/features/user/schema';
 import { protectedProcedure } from '@/server/orpc';
 
 const tags = ['user'];
